+++ CHANGES.md
@@ -2,6 +2,14 @@
 
 ## [Unreleased] - ReleaseDate
 
+### Fixed
+
+- `deduplicate_namespaces` had a bug. Attributes, unlike elements, can only be
+  in a namespace if they have an explicit prefix; there is no defaulting. This
+  means that a prefix definition for a duplicate namespace can only safely be
+  removed if this prefix doesn't overlap with a default namespace, or if the
+  prefix is not in use by any attribute. We now check for this.
+
 ## [0.13.3] - 2023-06-26
 
 ### Fixed
@@ -355,8 +363,8 @@
 Initial public release.
 
 <!-- next-url -->
-[Unreleased]: https://github.com/faassen/xot/compare/v0.13.3...HEAD
 
+[Unreleased]: https://github.com/faassen/xot/compare/v0.13.3...HEAD
 [0.13.3]: https://github.com/faassen/xot/compare/v0.13.2...v0.13.3
 [0.13.2]: https://github.com/faassen/xot/compare/v0.13.1...v0.13.2
 [unreleased]: https://github.com/faassen/xot/compare/v0.13.0...v0.13.1